+++ src/commonMain/kotlin/com/insanusmokrassar/krontab/internal/Ranges.kt
@@ -1,7 +1,0 @@
-package com.insanusmokrassar.krontab.internal
-
-internal val monthRange = 0 .. 11
-internal val dayOfMonthRange = 0 .. 30
-internal val hoursRange = 0 .. 23
-internal val minutesRange = 0 .. 59
-internal val secondsRange = minutesRange