+++ src/commonTest/kotlin/com/insanusmokrassar/krontab/utils/RunTest.kt
@@ -1,8 +1,0 @@
-package com.insanusmokrassar.krontab.utils
-
-import kotlinx.coroutines.CoroutineScope
-
-/**
- * Workaround to use suspending functions in unit tests
- */
-expect fun runTest(block: suspend (scope : CoroutineScope) -> Unit)