+++ src/jsTest/kotlin/com/insanusmokrassar/krontab/utils/RunTest.kt
@@ -1,5 +1,0 @@
-package com.insanusmokrassar.krontab.utils
-
-import kotlinx.coroutines.*
-
-actual fun runTest(block: suspend (scope : CoroutineScope) -> Unit): dynamic = GlobalScope.promise { block(this) }