+++ CHANGELOG.md
@@ -1,11 +1,23 @@
 # Changelog
 
-### 0.3.3
+## 0.4.0
+
+**BREAKING CHANGES**
+Package of project has been changed. Migration:
+
+* Replace in your dependencies `com.insanusmokrassar:krontab` by `dev.inmo:krontab`
+* Replace in your project all imports `com.insanusmokrassar.krontab` by `dev.inmo.krontab`
+
+* Versions:
+    * `Kotlin`: `1.4.10` -> `1.4.20`
+    * `Klock`: `1.12.1` -> `2.0.0`
+
+## 0.3.3
 
 * Versions:
     * `Coroutines`: `1.3.9` -> `1.4.1`
 
-### 0.3.2
+## 0.3.2
 
 * Function `TimeBuilder#each` was added (works as `at`)
 * Add opportunity to use `first` shortcuts:
@@ -15,7 +27,7 @@
     * Value property `TimeBuilder#last` for including via functions like `TimeBuilder#at`
     * Shortcut for kron string format `l` or `L`
 
-### 0.3.1
+## 0.3.1
 
 * Versions:
     * `Kotlin`: `1.4.0` -> `1.4.10`
@@ -37,7 +49,7 @@
     * Coroutines `1.3.7` -> `1.3.8`
 * Ranges support were included. Now it is possible to correctly use syntax `0-5` in strings schedules
 
-### 0.2.3
+## 0.2.3
 
 * Updates in libraries:
     * Kotlin `1.3.70` -> `1.3.72`
@@ -49,14 +61,14 @@
     * `KronScheduler#doWhile` was rewritten to use `KronScheduler#doOnce` for calculations of `block` result
 * New `buildSchedule(String)` function as a shortcut for `createSimpleScheduler(String)`
 
-### 0.2.2
+## 0.2.2
 
 * Updates in libraries:
     * Kotlin `1.3.61` -> `1.3.70`
     * Coroutines `1.3.3` -> `1.3.5`
     * Klock `1.8.6` -> `1.10.0`
 
-### 0.2.1
+## 0.2.1
 
 * Added support of flows: now any `KronScheduler` can be convert to `Flow<DateTime>` using `asFlow` extension